+++ nengo_ocl/raggedarray.py
@@ -2,9 +2,11 @@
 Numpy implementation of RaggedArray data structure.
 
 """
-from __future__ import print_function
 
-from nengo.utils.compat import is_iterable, StringIO
+from io import StringIO
+
+from nengo.utils.numpy import is_iterable
+
 import numpy as np
 
 from nengo_ocl.utils import round_up